cleanup: split out Interpolate code into separate file
And add copyright headers.
